update and fix bug in --with-exuberant-ctags option.

+++ common/gnu_global.sh
@@ -5,14 +5,23 @@ tarball=${global}.tar.gz
 
 url=http://tamacom.com/global/${tarball}
 
-yum install ctags python-pygments gcc ncurses-devel
-curl -O "${url}"
+echo "=> installing dependent packages..."
+yum install -y ctags python-pygments gcc ncurses-devel
+if [ ! -f ${taball} ]; then
+	echo "=> ${tarball} not found, fetching..."
+	curl -O "${url}"
+else
+	echo "=> found ${tarball}, use it..."
+fi
+echo "=> building ${global}..."
 tar zxvf ${tarball}
 cd ${global}
-./configure --prefix=/usr/local --with-exuberant-ctags 2>&1 | tee log.configure
+./configure --prefix=/usr/local --with-exuberant-ctags=/usr/bin/ctags 2>&1 | tee log.configure
 make 2>&1 | tee log.make
+echo "=> installing ${global}..."
 sudo make install 2>&1 | tee log.install
 
+echo "=> setting global..."
 #cp /usr/local/share/gtags/gtags.conf /usr/local/etc/
 sed -e 's/:tc=native:/\0tc=pygments:/' /usr/local/share/gtags/gtags.conf > /usr/local/etc/gtags.conf
